[BIT-21] Refactor Wizard
