+++ resources/js/app.js
@@ -1,8 +1,8 @@
-import { createApp, h } from 'vue'
-import { createInertiaApp } from '@inertiajs/inertia-vue3'
-import { resolvePageComponent } from "laravel-vite-plugin/inertia-helpers";
+import {createApp, h} from 'vue'
+import {createInertiaApp} from '@inertiajs/inertia-vue3'
+import {resolvePageComponent} from "laravel-vite-plugin/inertia-helpers";
 import '../css/app.css';
-
+import FontAwesomeIcon from './icons.ts';
 
 createInertiaApp({
     resolve: (name) =>
@@ -10,8 +10,9 @@ createInertiaApp({
             `./Pages/${name}.vue`,
             import.meta.glob("./Pages/**/*.vue")
         ),
-    setup({ el, app, props, plugin }) {
-        return createApp({ render: () => h(app, props) })
+    setup({el, app, props, plugin}) {
+        return createApp({render: () => h(app, props)})
+            .component('font-awesome-icon', FontAwesomeIcon)
             .use(plugin)
             .mount(el);
     },
