+++ resources/js/icons.ts
@@ -1,0 +1,14 @@
+import {library} from '@fortawesome/fontawesome-svg-core';
+import {FontAwesomeIcon} from "@fortawesome/vue-fontawesome";
+
+
+// Icon Imports
+import {faCircleInfo,} from "@fortawesome/free-solid-svg-icons";
+
+// Add Icons to Library
+library.add(
+    faCircleInfo,
+);
+
+
+export default FontAwesomeIcon;
